requirements.txt
customtkinter==5.2.2
pillow==11.1.0